Reorganized match files
