+++ app/controllers/matches_controller.rb
@@ -1,4 +1,5 @@
 class MatchesController < ApplicationController
+  include MatchValidation
   def index
     active_tab
     @matches = Match.all
@@ -21,19 +22,19 @@ class MatchesController < ApplicationController
   def create
     @match = Match.new(redGoal: match_params[:redGoal], blueGoal: match_params[:blueGoal], winner: match_params[:winner])
 
-    if validation.empty?
+    if validate(match_params).empty?
       update_players_for_match
       @match.save
       redirect_to matches_path
     else
-      redirect_to new_match_path(validation)
+      redirect_to new_match_path(validate(match_params))
     end
   end
 
   def update
     @match = Match.find(params[:id])
 
-    if validation.empty?
+    if validate(match_params).empty?
       update_hash = {
           redGoal: match_params[:redGoal],
           blueGoal: match_params[:blueGoal],
@@ -43,7 +44,7 @@ class MatchesController < ApplicationController
       @match.update(update_hash)
       redirect_to matches_path
     else
-      redirect_to action: :edit, id: @match.id, params: validation
+      redirect_to action: :edit, id: @match.id, params: validate(match_params)
     end
   end
 
@@ -66,48 +67,6 @@ class MatchesController < ApplicationController
 
   def update_players_for_match
     @match.add_players_for_match(match_params[:redAtt], match_params[:redDef], match_params[:blueAtt], match_params[:blueDef])
-  end
-
-  def validation
-    error = {}
-    @goal_check = false
-    @player_check = false
-    @blanks = []
-
-    params_list = %w[redAtt redDef blueAtt blueDef redGoal blueGoal]
-    params_list.each { |p| check_selected(match_params[p]) }
-
-    if @blanks.count(true) == 6
-      check_goal(match_params[:redGoal].to_i, match_params[:blueGoal].to_i)
-      check_players([match_params[:redAtt], match_params[:redDef], match_params[:blueAtt], match_params[:blueDef]])
-      error[:goal] = false if @goal_check
-      error[:player] = false if @player_check
-    end
-
-    error[:blanks] = @blanks if @blanks.count(true) < 6
-    error
-  end
-
-  def check_selected (field)
-    field.present? ? @blanks.push(true) : @blanks.push(false)
-  end
-
-  def check_goal (goal1, goal2)
-    @goal_check = true if goal1 == 10 && goal2 == 10
-    @goal_check = true if goal1 < 10 && goal2 < 10
-
-    if goal1 == 'Score'|| goal2 == 'Score'
-      @goal_check = true
-    end
-  end
-
-  def check_players (players)
-    players.each_with_index do |p|
-      other_players = players.clone
-      other_players.delete_at(players.index(p))
-      @player_check = other_players.include?(p)
-      break if @player_check
-    end
   end
 
   def check_for_error_params
